chore: fixing up native endpoints (#23)
