+++ .github/workflows/deploy-staging.yml
@@ -52,7 +52,7 @@ jobs:
         # WASM reproducibility check akin to SourceScan won't be available for staging contracts, deployed from PRs
         run: |
           if [ -n "$NEWLY_CREATED" ]; then
-            INIT_ARGS=$(cargo run --package test-utils --example generate_testnet_configuration)
+            INIT_ARGS=$(cargo run --package contract-mvp --example generate_testnet_configuration)
             cargo near deploy build-reproducible-wasm --skip-git-remote-check "${{ env.NEAR_CONTRACT_PR_STAGING_ACCOUNT_ID }}" \
               with-init-call new \
                 json-args "$INIT_ARGS" \
