+++ Cargo.toml
@@ -71,3 +71,6 @@ overflow-checks = true
 
 [lints]
 workspace = true
+
+[[example]]
+name = "generate_testnet_configuration"
