+++ test-utils/Cargo.toml
@@ -8,6 +8,3 @@ near-sdk.workspace = true
 near-workspaces.workspace = true
 templar-common.workspace = true
 tokio.workspace = true
-
-[[example]]
-name = "generate_testnet_configuration"
