+++ test-utils/src/lib.rs
@@ -31,32 +31,55 @@ pub const COLLATERAL_HALF_PRICE: OraclePriceProof = OraclePriceProof {
     borrow_asset_price: Rational::<u128>::one(),
 };
 
+pub enum TestAsset {
+    Native,
+    Nep141(Contract),
+}
+
+impl TestAsset {
+    pub fn is_native(&self) -> bool {
+        matches!(self, Self::Native)
+    }
+
+    pub fn nep141_id(&self) -> Option<&AccountId> {
+        if let Self::Nep141(ref contract) = self {
+            Some(contract.id())
+        } else {
+            None
+        }
+    }
+}
+
 pub struct TestController {
     pub worker: Worker<Sandbox>,
     pub contract: Contract,
-    pub borrow_asset: Contract,
-    pub collateral_asset: Contract,
+    pub borrow_asset: TestAsset,
+    pub collateral_asset: TestAsset,
 }
 
 impl TestController {
     pub async fn storage_deposits(&self, account: &Account) {
         println!("Performing storage deposits for {}...", account.id());
-        account
-            .call(self.borrow_asset.id(), "storage_deposit")
-            .args_json(json!({}))
-            .deposit(NearToken::from_near(1))
-            .transact()
-            .await
-            .unwrap()
-            .unwrap();
-        account
-            .call(self.collateral_asset.id(), "storage_deposit")
-            .args_json(json!({}))
-            .deposit(NearToken::from_near(1))
-            .transact()
-            .await
-            .unwrap()
-            .unwrap();
+        if let TestAsset::Nep141(ref borrow_asset) = self.borrow_asset {
+            account
+                .call(borrow_asset.id(), "storage_deposit")
+                .args_json(json!({}))
+                .deposit(NearToken::from_near(1))
+                .transact()
+                .await
+                .unwrap()
+                .unwrap();
+        }
+        if let TestAsset::Nep141(ref collateral_asset) = self.collateral_asset {
+            account
+                .call(collateral_asset.id(), "storage_deposit")
+                .args_json(json!({}))
+                .deposit(NearToken::from_near(1))
+                .transact()
+                .await
+                .unwrap()
+                .unwrap();
+        }
     }
 
     pub async fn get_configuration(&self) -> MarketConfiguration {
@@ -69,24 +92,34 @@ impl TestController {
             .unwrap()
     }
 
+    pub async fn supply_native(&self, supply_user: &Account, amount: u128) {
+        supply_user
+            .call(self.contract.id(), "supply_native")
+            .args_json(json!({}))
+            .deposit(NearToken::from_yoctonear(amount))
+            .transact()
+            .await
+            .unwrap()
+            .unwrap();
+    }
+
     pub async fn supply(&self, supply_user: &Account, amount: u128) {
         println!(
             "{} transferring {amount} tokens for supply...",
             supply_user.id()
         );
-        supply_user
-            .call(self.borrow_asset.id(), "ft_transfer_call")
-            .args_json(json!({
-                "receiver_id": self.contract.id(),
-                "amount": U128(amount),
-                "msg": serde_json::to_string(&Nep141MarketDepositMessage::Supply).unwrap(),
-            }))
-            .deposit(NearToken::from_yoctonear(1))
-            .max_gas()
-            .transact()
-            .await
-            .unwrap()
-            .unwrap();
+        match self.borrow_asset {
+            TestAsset::Native => self.supply_native(supply_user, amount).await,
+            TestAsset::Nep141(_) => {
+                self.borrow_asset_transfer_call(
+                    supply_user,
+                    self.contract.id(),
+                    amount,
+                    &serde_json::to_string(&Nep141MarketDepositMessage::Supply).unwrap(),
+                )
+                .await;
+            }
+        }
     }
 
     pub async fn get_supply_position(&self, account_id: &AccountId) -> Option<SupplyPosition> {
@@ -111,24 +144,34 @@ impl TestController {
             .unwrap()
     }
 
+    pub async fn collateralize_native(&self, borrow_user: &Account, amount: u128) {
+        borrow_user
+            .call(self.contract.id(), "collateralize_native")
+            .args_json(json!({}))
+            .deposit(NearToken::from_yoctonear(amount))
+            .transact()
+            .await
+            .unwrap()
+            .unwrap();
+    }
+
     pub async fn collateralize(&self, borrow_user: &Account, amount: u128) {
         println!(
             "{} transferring {amount} tokens for collateral...",
             borrow_user.id(),
         );
-        borrow_user
-            .call(self.collateral_asset.id(), "ft_transfer_call")
-            .args_json(json!({
-                "receiver_id": self.contract.id(),
-                "amount": U128(amount),
-                "msg": serde_json::to_string(&Nep141MarketDepositMessage::Collateralize).unwrap(),
-            }))
-            .deposit(NearToken::from_yoctonear(1))
-            .max_gas()
-            .transact()
-            .await
-            .unwrap()
-            .unwrap();
+        match self.collateral_asset {
+            TestAsset::Native => self.collateralize_native(borrow_user, amount).await,
+            TestAsset::Nep141(_) => {
+                self.collateral_asset_transfer_call(
+                    borrow_user,
+                    self.contract.id(),
+                    amount,
+                    &serde_json::to_string(&Nep141MarketDepositMessage::Collateralize).unwrap(),
+                )
+                .await;
+            }
+        }
     }
 
     pub async fn get_borrow_position(&self, account_id: &AccountId) -> Option<BorrowPosition> {
@@ -186,29 +229,49 @@ impl TestController {
     }
 
     pub async fn collateral_asset_balance_of(&self, account_id: &AccountId) -> u128 {
-        self.collateral_asset
-            .view("ft_balance_of")
-            .args_json(json!({
-                "account_id": account_id,
-            }))
-            .await
-            .unwrap()
-            .json::<U128>()
-            .unwrap()
-            .0
+        match self.collateral_asset {
+            TestAsset::Native => self
+                .worker
+                .view_account(self.contract.id())
+                .await
+                .map(|v| v.balance.as_yoctonear())
+                .unwrap(),
+            TestAsset::Nep141(ref collateral_asset) => {
+                collateral_asset
+                    .view("ft_balance_of")
+                    .args_json(json!({
+                        "account_id": account_id,
+                    }))
+                    .await
+                    .unwrap()
+                    .json::<U128>()
+                    .unwrap()
+                    .0
+            }
+        }
     }
 
     pub async fn borrow_asset_balance_of(&self, account_id: &AccountId) -> u128 {
-        self.borrow_asset
-            .view("ft_balance_of")
-            .args_json(json!({
-                "account_id": account_id,
-            }))
-            .await
-            .unwrap()
-            .json::<U128>()
-            .unwrap()
-            .0
+        match self.borrow_asset {
+            TestAsset::Native => self
+                .worker
+                .view_account(self.contract.id())
+                .await
+                .map(|v| v.balance.as_yoctonear() - v.locked.as_yoctonear())
+                .unwrap(),
+            TestAsset::Nep141(ref borrow_asset) => {
+                borrow_asset
+                    .view("ft_balance_of")
+                    .args_json(json!({
+                        "account_id": account_id,
+                    }))
+                    .await
+                    .unwrap()
+                    .json::<U128>()
+                    .unwrap()
+                    .0
+            }
+        }
     }
 
     pub async fn asset_transfer(
@@ -268,8 +331,19 @@ impl TestController {
         receiver_id: &AccountId,
         amount: u128,
     ) {
-        self.asset_transfer(self.borrow_asset.id(), sender, receiver_id, amount)
-            .await;
+        match self.borrow_asset {
+            TestAsset::Native => {
+                sender
+                    .transfer_near(receiver_id, NearToken::from_yoctonear(amount))
+                    .await
+                    .unwrap()
+                    .unwrap();
+            }
+            TestAsset::Nep141(ref contract) => {
+                self.asset_transfer(contract.id(), sender, receiver_id, amount)
+                    .await;
+            }
+        }
     }
 
     pub async fn borrow_asset_transfer_call(
@@ -279,25 +353,54 @@ impl TestController {
         amount: u128,
         msg: &str,
     ) -> ExecutionSuccess {
-        self.asset_transfer_call(self.borrow_asset.id(), sender, receiver_id, amount, msg)
-            .await
+        if let TestAsset::Nep141(ref borrow_asset) = self.borrow_asset {
+            self.asset_transfer_call(borrow_asset.id(), sender, receiver_id, amount, msg)
+                .await
+        } else {
+            panic!("Cannot perform an ft_transfer_call with a native asset");
+        }
     }
 
-    pub async fn repay(&self, borrow_user: &Account, amount: u128) {
-        println!("{} repaying {amount} tokens...", borrow_user.id());
+    pub async fn collateral_asset_transfer_call(
+        &self,
+        sender: &Account,
+        receiver_id: &AccountId,
+        amount: u128,
+        msg: &str,
+    ) -> ExecutionSuccess {
+        if let TestAsset::Nep141(ref collateral_asset) = self.collateral_asset {
+            self.asset_transfer_call(collateral_asset.id(), sender, receiver_id, amount, msg)
+                .await
+        } else {
+            panic!("Cannot perform an ft_transfer_call with a native asset");
+        }
+    }
+
+    pub async fn repay_native(&self, borrow_user: &Account, amount: u128) {
         borrow_user
-            .call(self.borrow_asset.id(), "ft_transfer_call")
-            .args_json(json!({
-                "receiver_id": self.contract.id(),
-                "amount": U128(amount),
-                "msg": serde_json::to_string(&Nep141MarketDepositMessage::Repay).unwrap(),
-            }))
-            .deposit(NearToken::from_yoctonear(1))
-            .max_gas()
+            .call(self.contract.id(), "repay_native")
+            .args_json(json!({}))
+            .deposit(NearToken::from_yoctonear(amount))
             .transact()
             .await
             .unwrap()
             .unwrap();
+    }
+
+    pub async fn repay(&self, borrow_user: &Account, amount: u128) {
+        println!("{} repaying {amount} tokens...", borrow_user.id());
+        match self.borrow_asset {
+            TestAsset::Native => self.repay_native(borrow_user, amount).await,
+            TestAsset::Nep141(_) => {
+                self.borrow_asset_transfer_call(
+                    borrow_user,
+                    self.contract.id(),
+                    amount,
+                    &serde_json::to_string(&Nep141MarketDepositMessage::Repay).unwrap(),
+                )
+                .await;
+            }
+        }
     }
 
     pub async fn harvest_yield(&self, supply_user: &Account) {
@@ -330,7 +433,11 @@ impl TestController {
             .unwrap();
     }
 
-    pub async fn withdraw_supply_yield(&self, supply_user: &Account, amount: Option<u128>) {
+    pub async fn withdraw_supply_yield(
+        &self,
+        supply_user: &Account,
+        amount: Option<u128>,
+    ) -> ExecutionSuccess {
         println!("{} withdrawing supply yield...", supply_user.id());
         supply_user
             .call(self.contract.id(), "withdraw_supply_yield")
@@ -340,7 +447,7 @@ impl TestController {
             .transact()
             .await
             .unwrap()
-            .unwrap();
+            .unwrap()
     }
 
     pub async fn get_static_yield(&self, account_id: &AccountId) -> Option<StaticYieldRecord> {
@@ -360,7 +467,7 @@ impl TestController {
         borrow_user: &Account,
         amount: u128,
         price: Option<OraclePriceProof>,
-    ) {
+    ) -> ExecutionSuccess {
         println!("{} withdrawing {amount} collateral...", borrow_user.id());
         borrow_user
             .call(self.contract.id(), "withdraw_collateral")
@@ -371,7 +478,7 @@ impl TestController {
             .transact()
             .await
             .unwrap()
-            .unwrap();
+            .unwrap()
     }
 
     pub async fn create_supply_withdrawal_request(&self, supply_user: &Account, amount: u128) {
@@ -429,6 +536,26 @@ impl TestController {
             .unwrap();
     }
 
+    pub async fn liquidate_native(
+        &self,
+        liquidator_user: &Account,
+        account_id: &AccountId,
+        borrow_asset_amount: u128,
+        oracle_price_proof: OraclePriceProof,
+    ) {
+        liquidator_user
+            .call(self.contract.id(), "liquidate_native")
+            .args_json(json!({
+                "account_id": account_id,
+                "oracle_price_proof": oracle_price_proof,
+            }))
+            .deposit(NearToken::from_yoctonear(borrow_asset_amount))
+            .transact()
+            .await
+            .unwrap()
+            .unwrap();
+    }
+
     pub async fn liquidate(
         &self,
         liquidator_user: &Account,
@@ -442,17 +569,30 @@ impl TestController {
             account_id,
             borrow_asset_amount,
         );
-        self.borrow_asset_transfer_call(
-            liquidator_user,
-            self.contract.id(),
-            borrow_asset_amount,
-            &serde_json::to_string(&Nep141MarketDepositMessage::Liquidate(LiquidateMsg {
-                account_id: account_id.clone(),
-                oracle_price_proof,
-            }))
-            .unwrap(),
-        )
-        .await;
+        match self.borrow_asset {
+            TestAsset::Native => {
+                self.liquidate_native(
+                    liquidator_user,
+                    account_id,
+                    borrow_asset_amount,
+                    oracle_price_proof,
+                )
+                .await
+            }
+            TestAsset::Nep141(_) => {
+                self.borrow_asset_transfer_call(
+                    liquidator_user,
+                    self.contract.id(),
+                    borrow_asset_amount,
+                    &serde_json::to_string(&Nep141MarketDepositMessage::Liquidate(LiquidateMsg {
+                        account_id: account_id.clone(),
+                        oracle_price_proof,
+                    }))
+                    .unwrap(),
+                )
+                .await;
+            }
+        }
     }
 
     #[allow(unused)] // This is useful for debugging tests
@@ -535,7 +675,7 @@ pub static WASM_MOCK_FT: OnceCell<Vec<u8>> = OnceCell::const_new();
 
 pub async fn setup_market(
     worker: &Worker<Sandbox>,
-    configuration: MarketConfiguration,
+    configuration: &MarketConfiguration,
 ) -> Contract {
     let wasm = WASM_MARKET
         .get_or_init(|| async { near_workspaces::compile_project("./").await.unwrap() })
@@ -619,8 +759,9 @@ pub async fn setup_everything(
             .with_static(insurance_yield_user.id().clone(), 1),
     );
     customize_market_configuration(&mut config);
+
     let (contract, borrow_asset, collateral_asset) = tokio::join!(
-        setup_market(&worker, config),
+        setup_market(&worker, &config),
         deploy_ft(
             borrow_asset,
             "Borrow Asset",
@@ -636,6 +777,15 @@ pub async fn setup_everything(
             100000,
         ),
     );
+
+    let collateral_asset = config
+        .collateral_asset
+        .into_nep141()
+        .map_or(TestAsset::Native, |_| TestAsset::Nep141(collateral_asset));
+    let borrow_asset = config
+        .borrow_asset
+        .into_nep141()
+        .map_or(TestAsset::Native, |_| TestAsset::Nep141(borrow_asset));
 
     let c = TestController {
         worker,
